refactor(whispering): clarify overlay runtime ownership
Keep the hosted SPA out of the overlay runtime registry and isolate native window mechanics behind the Tauri platform graph.
