+++ apps/whispering/src/routes/(app)/_runtime/attach-recording-overlay.browser.ts
@@ -1,4 +1,4 @@
-/** Browser renders the recording pill in-page, so it owns no overlay window. */
-export function attachRecordingOverlay() {
-	return () => {};
-}
+import type { RuntimeOwner } from './types';
+
+/** Browser renders the recording pill in-page, so it owns no overlay runtime. */
+export const recordingOverlayRuntimeOwner: RuntimeOwner | null = null;
